+++ src/ConfirmDialog.js
@@ -1,30 +1,30 @@
 import {
+  Box,
   Button,
+  Checkbox,
+  CircularProgress,
   Dialog,
   DialogActions,
   DialogContent,
   DialogContentText,
   DialogTitle,
+  FormControlLabel,
+  Typography,
+  useMediaQuery,
   useTheme,
-  CircularProgress
 } from '@material-ui/core';
 import produce from 'immer';
 import { forwardRef, useImperativeHandle, useRef, useState } from 'react';
 import ReCAPTCHA from 'react-google-recaptcha';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import { useConfig } from './api';
-
-const preventDefault = (event) => {
-  event.preventDefault();
-};
-
-export const isCaptchaRequired = (config) =>
-  config.publicForm &&
-  // Form may not have `requireCaptcha` defined.
-  (config.requireCaptcha || config.requireCaptcha === undefined);
+import { hasCaptcha, hasConsent, hasPreview } from './customizations';
+import { PrintButton } from './PrintButton';
+import Field from './Review/Field';
+import { Customize } from './utils';
 
 export default forwardRef(function ConfirmDialog(
-  { title, description, children, onConfirm, ...other },
+  { title, description, children, onConfirm, confirmComponent, ...otherProps },
   ref
 ) {
   useImperativeHandle(ref, () => ({
@@ -32,7 +32,7 @@ export default forwardRef(function ConfirmDialog(
       return new Promise((resolve) => {
         confirmRef.current = function (confirm) {
           resolve(
-            isCaptchaRequired(other.config)
+            hasCaptcha(otherProps)
               ? confirm &&
                   produce(args, ([data]) => {
                     data.captchaChallenge = confirm;
@@ -51,24 +51,27 @@ export default forwardRef(function ConfirmDialog(
     },
     error() {
       handleError();
-    }
+    },
   }));
 
+  // const dialogContentFlexNoneClasses = useStylesDialogContentFlexNone();
+
   const confirmRef = useRef();
+  const hasCaptchaValue = hasCaptcha(otherProps);
+  const hasConsentValue = hasConsent(otherProps);
 
   const [open, setOpen] = useState(false);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(false);
-  const [showReCAPTCHA, setShowReCAPTCHA] = useState(
-    isCaptchaRequired(other.config)
-  );
+  const [showReCAPTCHA, setShowReCAPTCHA] = useState(hasCaptchaValue);
   const [captchaChallenge, setCaptchaChallenge] = useState();
+  const [consent, setConsent] = useState(false);
 
   function handleClose() {
     setLoading(false);
     setError(false);
     setOpen(false);
-    setShowReCAPTCHA(isCaptchaRequired(other.config));
+    setShowReCAPTCHA(hasCaptchaValue);
   }
 
   function handleDismiss() {
@@ -90,55 +93,96 @@ export default forwardRef(function ConfirmDialog(
   }
 
   function handleConfirm() {
-    confirmRef.current(
-      isCaptchaRequired(other.config) ? captchaChallenge : true
-    );
+    confirmRef.current(hasCaptchaValue ? captchaChallenge : true);
   }
 
+  const intl = useIntl();
+
   return (
-    <Dialog
-      open={open}
-      onClose={handleDismiss}
-      onClick={preventDefault}
-      aria-labelledby="confirm-dialog-title"
-      aria-describedby="confirm-dialog-description"
-    >
-      <DialogTitle id="confirm-dialog-title">{title}</DialogTitle>
-      {description && (
+    <Customize customizer={confirmComponent} {...otherProps}>
+      <Dialog
+        scroll="body"
+        open={open}
+        onClose={handleDismiss}
+        aria-labelledby="confirm-dialog-title"
+        aria-describedby="confirm-dialog-description"
+        fullScreen={useMediaQuery('print')}
+      >
+        <DialogTitle id="confirm-dialog-title">{title}</DialogTitle>
         <DialogContent>
-          <DialogContentText id="confirm-dialog-description">
-            {description}
-          </DialogContentText>
+          {description && (
+            <DialogContentText id="confirm-dialog-description">
+              {description}
+            </DialogContentText>
+          )}
+          {hasPreview(otherProps) && (
+            <>
+              <Typography variant="subtitle1" component="h3">
+                <FormattedMessage defaultMessage="Lähettävien tietojen esikatselu" />
+              </Typography>
+              <Field
+                preview // For selecting optional preview / review customization
+                root
+                {...otherProps}
+              />
+            </>
+          )}
+          {hasConsentValue && (
+            <FormControlLabel
+              key="consent"
+              control={
+                <Checkbox
+                  checked={consent}
+                  onChange={(event) => {
+                    setConsent(event.target.checked);
+                  }}
+                  name="consent"
+                  color="secondary"
+                />
+              }
+              label={intl.formatMessage({
+                defaultMessage: 'Vahvistan ilmoitetut tiedot oikeiksi',
+              })}
+            />
+          )}
+          {showReCAPTCHA && !loading && !error ? (
+            <DialogContentReCAPTCHA onChange={setCaptchaChallenge} />
+          ) : null}
+          {loading && (
+            <div>
+              <CircularProgress />
+            </div>
+          )}
+          {error && (
+            <DialogContentText>
+              <FormattedMessage defaultMessage="Lomakkeen lähetys ei onnistunut." />
+            </DialogContentText>
+          )}
         </DialogContent>
-      )}
-      { (showReCAPTCHA && !loading && !error) ? (
-        <DialogContentReCAPTCHA onChange={setCaptchaChallenge} />
-      ) : null}
-      {loading && (
-        <DialogContent>
-          <CircularProgress />
-        </DialogContent>
-      )}
-      {error && (
-        <DialogContent>
-          <FormattedMessage defaultMessage="Lomakkeen lähetys ei onnistunut." />
-        </DialogContent>
-      )}
-      <DialogActions>
-        <Button disabled={loading && !error} onClick={handleDismiss}>
-          <FormattedMessage defaultMessage="Peruuta" />
-        </Button>
-        <Button
-          disabled={(showReCAPTCHA && !captchaChallenge) || loading || error}
-          onClick={handleConfirm}
-          variant="contained"
-          color="primary"
-          autoFocus
-        >
-          <FormattedMessage defaultMessage="Lähetä" />
-        </Button>
-      </DialogActions>
-    </Dialog>
+        <Box displayPrint="none">
+          <DialogActions>
+            <Button disabled={loading && !error} onClick={handleDismiss}>
+              <FormattedMessage defaultMessage="Peruuta" />
+            </Button>
+            {hasPreview(otherProps) && <PrintButton />}
+            <Button
+              disabled={
+                (hasConsentValue && !consent) ||
+                (showReCAPTCHA && !captchaChallenge) ||
+                loading ||
+                error
+              }
+              onClick={handleConfirm}
+              variant="contained"
+              color="primary"
+              autoFocus
+            >
+              <FormattedMessage defaultMessage="Lähetä" />
+            </Button>
+          </DialogActions>
+        </Box>
+      </Dialog>
+    </Customize>
   );
 });