+++ src/Review/index.js
@@ -1,18 +1,17 @@
-import Button from '@material-ui/core/Button';
 import { makeStyles } from '@material-ui/core/styles';
-import { Print } from '@material-ui/icons';
-import { FormattedMessage, useIntl } from 'react-intl';
-import Markdown from '../Markdown';
-import Field from './Field';
 import { Alert } from '@material-ui/lab';
+import { useIntl } from 'react-intl';
+import Markdown from '../Markdown';
+import { PrintButton } from '../PrintButton';
+import Field from './Field';
 
 const useStyles = makeStyles((theme) => ({
   buttonContainer: {
     '& > *': {
       margin: theme.spacing(1),
-      '@media print' : {
-        display: 'none'
-      }
+      '@media print': {
+        display: 'none',
+      },
     },
   },
 }));
@@ -24,22 +23,13 @@ export default function Review(props) {
   return (
     <>
       <Alert severity="success">
-        { intl.formatMessage({defaultMessage: 'Lomake lähetetty!'}) }
+        {intl.formatMessage({ defaultMessage: 'Lomake lähetetty!' })}
       </Alert>
       <Field root {...props} />
       <Markdown>{props.config.successText}</Markdown>
       <div className={classes.buttonContainer}>
         {props.reviewProps?.actions}
-        <Button
-          variant="contained"
-          startIcon={<Print />}
-          onClick={global.print}
-          aria-label={intl.formatMessage({
-            defaultMessage: 'Tulosta. Avautuu PDF:nä uuteen ikkunaan.'
-          })}
-        >
-          <FormattedMessage defaultMessage="Tulosta" />
-        </Button>
+        <PrintButton />
       </div>
     </>
   );