+++ src/Review/types/array.js
@@ -1,10 +1,11 @@
 import { empty } from '..';
 import Field from '../Field';
 
-export default ({ formData, schema, uiSchema }) =>
+export default ({ formData, schema, uiSchema, ...otherProps }) =>
   formData?.length
     ? formData.map((formData, index) => (
         <Field
+          {...otherProps}
           key={index}
           schema={schema.items}
           uiSchema={uiSchema?.items}