+++ src/configToSchemas/types/select.js
@@ -1,5 +1,5 @@
 import { defineMessage } from 'react-intl';
-import { ensureValueIsAvailable } from '../../utils';
+import { ensureValueIsAvailable, sortChoices } from '../../utils';
 
 export default ({ config }) => {
   const {
@@ -12,20 +12,7 @@ export default ({ config }) => {
     autoSort = false,
     widget
   } = config;
-  const choicesSorted = [...choices];
-
-  if (autoSort) {
-    choicesSorted.sort((el1, el2) => {
-      const el1Name = el1.enumNames ? el1.enumNames : el1.enum;
-      const el2Name = el2.enumNames ? el2.enumNames : el2.enum;
-      if (el1Name > el2Name) {
-        return 1
-      }  else if (el1Name < el2Name) {
-        return -1
-      }
-      return 0;
-    })
-  }
+  const choicesSorted = sortChoices(choices, autoSort);
 
   return {
     schema: